feat: Warn user when "savedTokensPath" is not writable
Prevents saving the settings and warns the user to change the
"savedTokensPath" if it's detected that a token file cannot be created
in the provided location either due to permission error or any other
issues.

This will prevent problems later on when trying to save the fetched
tokens to the disk in the specified location.